prototypes/qwen-numpy/requirements.txt
numpy
llama-cpp-python
